adding and deleting project interface
